refactor: move FunctionsMerger into its own file. add first test
